requirements.txt
fastapi
uvicorn
sentence-transformers
spacy
scikit-learn
pydantic
python-multipart
pdfplumber
python-docx
supabase
numpy 